ml_model/requirements.txt
tensorflow
pandas
numpy
SQLAlchemy
mysql-connector-python
scikit-learn
python-dotenv
joblib
fastapi
uvicorn
